Add async io (#31)
* move constants outside of protocol

Constants need to live in a space to be used by all aspects of the
application

* move utils to higher level

* move hash out of protocol

* rename hash checksum/verify

* refactor hash and utils tests to live within same files

* move mailbox to data_structures dir

* add io module

* add test files and attribution to TigerBeetle

* working on a reimplementation of node

The current implementation of node is really wonky and doesn't have a
clear flow to how I actually want the API to look

* add protocol version enum

I'm not sure why i never had this

* update protocol version

* add message queue

* rename queue to message_queue

* add benchmarks for MessageQueue.enqueue and MessageQueue.dequeue

* improve MessageQueue benchmarks

* add enqueueMany and dequeueMany

* make MessageQueue.dequeue return optional

* adding comments

* working on node io

* add node to test

* got basic message recv working

* made a very hasty queue.remove method

* adding a global message pool

* add note for assigned_map

* reorganize code and pull structures out of the node file

Things like the message bus and message pool don't need to be defined
inside of the node since there is no direct dependency

* make validation for messages

* remove prev ref from message

I don't think i need this because we are only using unidirectional
queues at the moment. Perhaps when I get to making an actual Queue
message this will become necessary

* adding some basic logic for pings

* working on accepting connection

* bump message.headers values

Changes headers_checksum from u64 to u128
Changes body_checksum from u64 to u128
Changes origin_id from u64 to u128

This was done for 1 reason, uuids are 128bits

* start refactor

* working on a zero alloc parser

* parser2 isn't better :(

I think i'm doing too much copying and it is becoming cumbersome

* trying to figure out how I can work with many connections and allocators

* working on the client/node connection

* working on client

* rename conn to self

* add inbox count log

* working on logs

* got the node to read way faster

* working on a client that can connect and ping and pong

* made a lot of progress on the client sending messages to the node

The client can now send messages to the node and the node can pass them
through a handler function. The issue is now that the client needs to
figure out how to deal with messages and so does the node

* this is real life

* fix memory leak bug

Pong was not being dereferenced or even cleaned up for testing

* working on a basic client

* working on testing the capacity of the node

* playing with numbers

* change up message bus to work slightly differently

Message bus now uses a shorter window

* playing with parser

trying to debug an issue where the parser get's stuck into a bad state.
I think the connection should simply be closed when we get an invalid
header checkshum vs trying again. This isn't really the answer I want
but it is simple

* add connection disconnect on bad message

* make an accept message which passes ids to the connectees

* making message_bus handle the connections

The issue is that it would be inefficient for the message bus to figure
out how traverse back to the node. also is just unnecessary cognitive
load

* made a bunch of progress and started some really simplified benchmarking to figure out bottlenecks

* got client receive pong but there is a bug

* ping and pong are working