create the initial draft and benchmarks for serialize/deserialize and compare against encode/decode
The first draft is slower. But by around 200ns which isn't much. I think
that I can optimize this by not doing a second pass on the data when
calculating the checksum. However, with the additional complication and
safety checks required for the extension_headers field, I think that I
may not be able to climb much higher.