Feature/implement frame and frame parser (#8)
* working on the frame and frameheaders

* working on hashing correctly

* remove hash.zig from plain src dir and put in lib

* add a Frame to wrap FrameHeaders and payload

This simplifies the toBytes and fromBytes functions as the checksum can
be attached after the payload so that you don't have to do any crazy
reordering or rereading of the buffer just to ensure the checksum is
calculated. Easy peasy.

* rename hash.zig to checksum.zig

This is mostly done to not confuse myself since

* add basic benchmarks for checksum

* add benchmarks for checksum

* change from crc32 to xxHash32

It is just faster for large messages with pretty similar performance on
small messages to crc